Small restructuring
- Changed Character.GetDirection return type to int
- Made MoveA/B protected to Element
- Created public MoveMouseOne/Two, can be changed in the future so that move can be chosen through index

+++ Assets/Scripts/Character/Character.cs
@@ -114,14 +114,14 @@ public abstract class Character : MonoBehaviour
         }
     }
 
-    public float GetDirectionX()
+    public int GetDirectionX()
     {
         if (noneXInput) { return 0; }
         else if (IsLookingRight) { return 1;}
         else {  return -1; }
     }
 
-    public float GetDirectionY()
+    public int GetDirectionY()
     {
         if (noneYInput) { return 0; }
         else if (IsLookingUp) { return 1; }
